requirements.txt
ipython
ipykernel
jinja2
jsonschema 
typing-extensions
numpy
pandas
scikit-learn
matplotlib
seaborn
xgboost
lightgbm
python-dotenv
exceptiongroup 
pygments 
stack_data
oauthlib
kagglehub
tensorflow[and-cuda]
imblearn
catboost
shap
tqdm
pymongo[srv]
pymongo
certifi
mlflow==2.2.2
dagshub
dvc